+++ client/ui/board.jsx
@@ -20,6 +20,12 @@ class Board extends Base {
     const name = this.state.model.name;
     this.refs.rename.show({old_name: name, name: name});
   }
+
+  download() {
+    this.api.export_url((url) => {
+      window.open(url, '_blank');
+    }); 
+  }
   
   render() {
     const board = this.state.model;
@@ -31,12 +37,20 @@ class Board extends Base {
     }
     document.title = board.name;
 
-    const extra_nav = (
-      <TooltipIconButton
-         icon="mode_edit" tooltip="Rename board"
+    const extra_nav = [
+      (<TooltipIconButton
+         icon="mode_edit" tooltip="Rename board" key="edit"
          onMouseUp={() => this.show_rename()}
-        />
-    );
+       />)
+    ];
+    if (board.user.admin) {
+      extra_nav.unshift(
+        <TooltipIconButton
+          icon="file_download" tooltip="Download board data" key="download"
+          onMouseUp={() => this.download()}
+          />
+      );
+    }
     
     return (
       <div className="kb-board">